fixe project structure
