--- src/app/_shared/event-bus.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { EventBusService } from './event-bus.service';
-
-describe('EventBusService', () => {
-  let service: EventBusService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(EventBusService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
