+++ src/app/app-routing.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { RegisterComponent } from './register/register.component';
-import { LoginComponent } from './login/login.component';
-import { HomeComponent } from './home/home.component';
-import { ProfileComponent } from './profile/profile.component';
-import { BoardUserComponent } from './board-user/board-user.component';
-import { BoardModeratorComponent } from './board-moderator/board-moderator.component';
-import { BoardAdminComponent } from './board-admin/board-admin.component';
+import { RegisterComponent } from './auth/register/register.component';
+import { LoginComponent } from './auth/login/login.component';
+import { HomeComponent } from './features/home/home.component';
+import { ProfileComponent } from './auth/profile/profile.component';
+import { BoardUserComponent } from './features/board-user/board-user.component';
+import { BoardModeratorComponent } from './features/board-moderator/board-moderator.component';
+import { BoardAdminComponent } from './features/board-admin/board-admin.component';
 
 const routes: Routes = [
   { path: 'home', component: HomeComponent },
