+++ src/app/app.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 import { Subscription } from 'rxjs';
-import { StorageService } from './_services/storage.service';
-import { AuthService } from './_services/auth.service';
-import { EventBusService } from './_shared/event-bus.service';
+import { StorageService } from './core/services/storage.service';
+import { AuthService } from './core/services/auth.service';
+import { EventBusService } from './shared/event-bus.service';
 
 @Component({
   selector: 'app-root',
