+++ src/app/auth/login/login.component.html
@@ -8,7 +8,7 @@
         Manage your health records, book appointments, and take control of your healthcare journey.
       </p>
       <div class="mt-4 rounded-full bg-white p-4">
-        <img src="./Swiss_pharmacy_logo.ico" alt="AdvancedMD Logo" class="w-32 h-32 rounded-full shadow-lg">
+        <img src="Swiss_pharmacy_logo.ico" alt="AdvancedMD Logo" class="w-32 h-32 rounded-full shadow-lg">
       </div>
     </div>
 
@@ -63,10 +63,10 @@
 
         <div class="form-group">
           <div *ngIf="isLoginFailed" class="text-sm text-red-500 mt-1" role="alert">
-            {{ errorMessage }} 
+            {{ errorMessage }}
           </div>
         </div>
-        
+
 
         <!-- Submit Button -->
         <div>
