+++ src/app/features/board-admin/board-admin.component.ts
@@ -1,18 +1,17 @@
 import { Component, OnInit } from '@angular/core';
-import { UserService } from '../_services/user.service';
+import { UserService } from '../../core/services/user.service';
 
 @Component({
-  selector: 'app-home',
-  templateUrl: './home.component.html',
-  styleUrls: ['./home.component.css']
+  selector: 'app-board-admin',
+  templateUrl: './board-admin.component.html',
 })
-export class HomeComponent implements OnInit {
+export class BoardAdminComponent implements OnInit {
   content?: string;
 
   constructor(private userService: UserService) { }
 
   ngOnInit(): void {
-    this.userService.getPublicContent().subscribe({
+    this.userService.getAdminBoard().subscribe({
       next: data => {
         this.content = data;
       },
