+++ src/app/features/board-moderator/board-moderator.component.ts
@@ -1,18 +1,17 @@
 import { Component, OnInit } from '@angular/core';
-import { UserService } from '../_services/user.service';
+import { UserService } from '../../core/services/user.service';
 
 @Component({
-  selector: 'app-board-admin',
-  templateUrl: './board-admin.component.html',
-  styleUrls: ['./board-admin.component.css']
+  selector: 'app-board-moderator',
+  templateUrl: './board-moderator.component.html',
 })
-export class BoardAdminComponent implements OnInit {
+export class BoardModeratorComponent implements OnInit {
   content?: string;
 
   constructor(private userService: UserService) { }
 
   ngOnInit(): void {
-    this.userService.getAdminBoard().subscribe({
+    this.userService.getModeratorBoard().subscribe({
       next: data => {
         this.content = data;
       },
