+++ src/app/features/board-user/board-user.component.ts
@@ -1,10 +1,9 @@
 import { Component, OnInit } from '@angular/core';
-import { UserService } from '../_services/user.service';
+import { UserService } from '../../core/services/user.service';
 
 @Component({
   selector: 'app-board-user',
   templateUrl: './board-user.component.html',
-  styleUrls: ['./board-user.component.css']
 })
 export class BoardUserComponent implements OnInit {
   content?: string;
