+++ src/app/features/home/home.component.ts
@@ -1,18 +1,17 @@
 import { Component, OnInit } from '@angular/core';
-import { UserService } from '../_services/user.service';
+import { UserService } from '../../core/services/user.service';
 
 @Component({
-  selector: 'app-board-moderator',
-  templateUrl: './board-moderator.component.html',
-  styleUrls: ['./board-moderator.component.css']
+  selector: 'app-home',
+  templateUrl: './home.component.html',
 })
-export class BoardModeratorComponent implements OnInit {
+export class HomeComponent implements OnInit {
   content?: string;
 
   constructor(private userService: UserService) { }
 
   ngOnInit(): void {
-    this.userService.getModeratorBoard().subscribe({
+    this.userService.getPublicContent().subscribe({
       next: data => {
         this.content = data;
       },
