+++ src/app/shared/event.class.ts
